Combine "area" and "event" menus into a single autocomplete menu
This allows us to remove all the custom JS that updates the
event menu in response to the area, and to avoid all the sticky
edge-cases you get when one menu is filtered by another menu.

For now this is a non-AJAX autocomplete, which makes the code
even simpler. There are a LOT of events, though, so it may
be necessary to ajaxify this eventually.