Issue #4: Require at least one workflow step
For now I'm only validating that the "steps" hidden
form element is non-empty. In the long run it'd be
more robust to, additionally, deserialize the JSON
and validate its structure. (This is something we
already do when submitting the modal form, but it
has to be re-done because the data was sent to the
browser and back again.)