When editing existing steps, give precedence to DB fields over JSON
The modal form for adding/editing steps, returns the step's data
as a list of JS objects, each with a "name" field and a "value"
field. When we store this into the DB, we take some of those pieces
of data and store them in specific database columns (name, description,
etc), and we take others and convert them into the keys and values
of an associative array, which is then serialized into JSON and stored
in a "data" column in the DB.

When preparing to edit an existing step, I convert that DB record
back into the modal form's JSON format; a list of objects with
"name" and "value" fields. This is a combination of data from
the deserialized "data" column, and the named columns. These two
sources are combined using the array_merge() function.

I'd placed the named column values first in array_merge(), and the
deserialized "data" values second, which meant that if, say,
"description" were present in the JSON and a DB column, the editing
screen would use the value from the JSON. But it's better to give
the values from the DB columns precedence, so I'm swapping
those two orders around.