Allow update of existing workflow & steps
The form submission code so far was only handling the creation of new
workflows, with new steps. This commit lets it also handle updates
to existing workflows and steps.

Note that when editing an existing workflow, a single submission of
the "edit workflow" form, may result in multiple workflow steps
being inserted, deleted, and/or updated.