Refactoring and simplifying JS code
A lot of the code around dealing with the modal
form for workflow steps, was complicated by the
fact that the steps were captured from the form
in the format returned by jQuery().serializeArray().
This format is a list of objects, each with a
"name" and "field" value. It's useful for serializing,
but for anything else, it is a lot less convenient than
a normal JS object or PHP associative array.

We were already storing this data in the DB as a normal
JS object (JSON-encoded). I've now updated the rest of
the PHP and JS code around the form so that it keeps the
steps in normal JS object format most of the time.

I also simplified the parameters used for rendering the
modal form fragment via tool_trigger_output_fragment_new_step_form,
and regularized the names of the params so that they're
consistent with names for their corresponding form
fields used elsewhere.