Add form fields for editing existing steps
Placing "steporder" into the step form allows us
to match up the submitted form with the step in
our list of steps, when the form is submitted. This
works for steps that aren't yet in the DB yet,
as well as steps that are.

Since we complete replace the serialized step
object in stepjson's list, with one newly serialized
from the form fields, we also need to add the step's
ID number, in order for edits to existing steps to
map correctly to their underlying DB objects.