Use the correct capability check
The code for rendering the workflow step forms, was checking
for the "moodle/course:managegroups" capability, which is
probably a holdover from the example code this was based on.

Changing it to "tool/trigger:manageworkflows", which is the
permission needed for the other parts of the edit pages.

Also removing an unused namespace import I'd accidentally added
to the top of the file at some point.