Remove top "move up" arrow and bottom "move down" arrow
I'm actually removing these by adding a "visibility:hidden" style
to them, and then removing it when I set up the click handlers.

I've decided to go ahead and do that to all of the step action
icons, to prevent a user from clicking on one before the click
handler is ready, in case this is loading slowly for some reason.