Update PHPunit tests to reflect changes in the code
I've simplified the structure of the JSON-encoded steps
in the form's stepjson field, and I've added an "id"
field to the records returned from the workflow_process::processjson()
method.

My refactoring of workflow_process::processjson() has also
resulted in the "data" field always being present on the
returned step records, even if the data is empty.