Distinguish between a filter quitting a workflow, and erroring out
As I understand it, the purpose of a "filter" step is to end the
execution of a workflow unless it means some particular filter.
This seems like it should mean that if the filter returns false
for its success flag, the workflow finishes.

The existing code, however, treated a false "success" flag like
it was an error, retrying the workflow again later.

This change allows the process to be marked as "complete" if
it exits early due to a filter. It also adds two separate statuses,
STATUS_FINISHED and STATUS_FINISHED_EARLY to make it easier to
tell, from the table records, whether a workflow finished early
or not.

And it adds an exception handler so that we can continue processing
further workflows if one of them throws an exception.