Shared functions for task data field names & substitutions
For now I'm using the term "datafield" to refer to all of the fields
of data available to steps in a workflow when the workflow is
executed. Specifically, this is the standard fields of the event,
the logextra fields of the event, the "other" fields of the event,
and any fields added to $stepresults by lookup steps.

All scalar values from these sources are combined into a single
associative array with regularized names as the keys.

Workflow authors can then refer to a specific field by name
when a step requires it (like the field from which to get the
user ID for the user lookup step).

Additionally, trigger steps like the email step, can have
"template" strings that have placeholders for datafields
to be inserted, e.g. {username}. These will be replaced
with the current value of the datafield by that name, when
the workflow is executed.