Display language strings for steps in step table
We were previously displaying the hardcoded steptype string, which doubles
as part of the namespace: actions/lookups/filters.

I've updated the code so that now it uses the corresponding human-readable
language string, and, where applicable, pulls it fresh from the system.
(There is still some cacheing while you're in the process of editing a
workflow, because I didn't want to implement a webservice just to get
these, or depend on a langstring naming scheme.)