Initial privacy API implementation
Github issue #16

For now, I've just hard-coded a list of privacy metadata based on the current
step classes in the plugin.

In the longer run, if we want to make this plugin more modular
and expand the number of steps, we should treat the steps like subplugins,
and have each one declare its own privacy metadata.