Remove call to user_get_user_details() in user lookup step
Github issue #18: The user lookup step was using user_get_user_details()
to pull really really full data about the user. But that function is
rather old and complicated, and it could take many different paths
depending on the user's details; and some of those paths had dependencies
on non-autoloaded Moodle libraries being loaded in. So I'm replacing
it with the simpler method of only returning the fields from
the mdl_user table, plus the results of calling fullname().

Since we're no longer using user_get_user_details(), we can also add
support for displaying data about users with the "deleted" flag (which
could be very handy for user deletion events.) One of the dependencies
of user_get_user_details() would complain when trying to process a deleted
user, but that's no longer a concern.