Remove empty test method
I accidentally left in place the empty test method in which I was going
to write the test for the "nodeleted" setting in the user lookup step.
I wound up instead combining that test into the existing test for the
situation where the user has the "deleted" flag, but I forgot to
remove this empty skeleton.