More things for privacy provider to delete
The trait I've used for the privacy provider will delete all the events relating
to a user, when deletion is requested. But we also have the tool_trigger_queue.eventid
column, which is a foreign key to tool_trigger_events. So we should also
delete records from tool_trigger_queue when their associated event is deleted.
This is not so much a privacy concern (because the queue table doesn't store any data
about the events), but it will probably help prevent errors from incomplete data.