Use correct workflow queue status constants in cleanup task
Github issue #21. The cleanup task had SQL queries that were
based on an older version of the queue status constants.

While I was at it, I rewrote the event cleanup query to use
a NOT EXISTS rather than a NOT IN subquery, because NOT EXISTS
is often more efficient (although increasingly, database planning
engines are smart enough to automatically convert a NOT IN to a
NOT EXISTS, so it doesn't always make a difference anymore.)