Add "personal data" and "send spam" to risk profile of capabilities
Github Issue #22. You can use the plugin to view users' personal data,
and to send out messages. So in addition to the RISK_CONFIG flag,
it also needs RISK_PERSONAL and RISK_SPAM.

(These risks are based on the current set of steps bundled with the plugin.
If we implement additional steps, or add an API for other plugins to
provide steps, we may need to add the other risk flags as well.)