Fixed issue with duplicate emails
If there is a case multiple user accounts exist in the installation with the same e-mail address (I know, this isn't recommended, but the use case requires this), this will fail if you only do a lookup based upon email. You should use the id if this is present based upon earlier lookups.